trash/ai_bail_agent1/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1

# LLM providers (optional - install what you need)
openai>=1.0.0
anthropic>=0.7.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# For memory and persistence
jsonschema>=4.17.0

# Optional for better performance
numpy>=1.24.0
pandas>=2.0.0

# For development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
ipython>=8.0.0

# For potential future features (PDF/DOCX generation)
# reportlab>=4.0.0
# python-docx>=1.0.0
# pypdf>=3.0.0